# Conflicts:
#	source/com.kms.katalon.application/src/com/kms/katalon/feature/KSEFeature.java
#	source/com.kms.katalon.composer.execution/src/com/kms/katalon/composer/execution/settings/ExecutionSettingPage.java
#	source/com.kms.katalon.execution/src/com/kms/katalon/execution/configuration/AbstractRunConfiguration.java
#	source/com.kms.katalon.execution/src/com/kms/katalon/execution/util/ExecutionUtil.java

+++ source/com.kms.katalon.dal.file/src/com/kms/katalon/dal/fileservice/manager/EntityFileServiceManager.java
@@ -259,12 +259,14 @@ public class EntityFileServiceManager {
 
     public static void deleteFolder(FolderEntity folder) throws Exception {
         if (folder != null) {
-            List<FileEntity> childEntities = getChildren(folder, FileEntity.class);
-            for (FileEntity childEntity : childEntities) {
-                if (childEntity instanceof FolderEntity) {
-                    deleteFolder((FolderEntity) childEntity);
-                } else {
-                    delete(childEntity);
+            if (folder.getFolderType() != FolderType.USER) {
+                List<FileEntity> childEntities = getChildren(folder, FileEntity.class);
+                for (FileEntity childEntity : childEntities) {
+                    if (childEntity instanceof FolderEntity) {
+                        deleteFolder((FolderEntity) childEntity);
+                    } else {
+                        delete(childEntity);
+                    }
                 }
             }
             EntityService.getInstance().deleteEntity(folder);
